ldsp-uni-backend/pyproject.toml
[project]
name = "ldsp-uni-backend"
version = "0.1.0"
description = "LDStatusPro unified AI admin backend (FastAPI)"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy[asyncio]>=2.0.29",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.2.0",
    "chromadb>=0.4.24",
    "httpx>=0.27.0",
    "apscheduler>=3.10.0",
    "structlog>=24.1.0",
    "tiktoken>=0.6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["data*", "tests*", "migrations*", "scripts*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
